+++ src/modules/enspostprocess_mod.f90
@@ -45,6 +45,8 @@ module ensPostProcess_mod
   ! public procedures
   public :: epp_postProcess
 
+  integer, parameter        :: maxNumLevels = 200
+  
 contains
 
   !----------------------------------------------------------------------
@@ -105,7 +107,7 @@ contains
     logical  :: huLimitsBeforeRecenter   ! Choose to apply humidity limits before recentering
     logical  :: imposeSaturationLimit  ! switch for choosing to impose saturation limit of humidity
     logical  :: imposeRttovHuLimits    ! switch for choosing to impose the RTTOV limits on humidity
-    real(8)  :: weightRecenter         ! weight applied to recentering increment
+    real(8)  :: weightRecenter(maxNumLevels)  ! weight applied to recentering increment (between 0 and 1; 0 means no recentering)
     real(8)  :: weightRecenterLand     ! weight applied to recentering increment for land variables
     integer  :: numMembersToRecenter   ! number of members that get recentered on supplied analysis
     logical  :: useOptionTableRecenter ! use values in the optiontable file
@@ -165,7 +167,7 @@ contains
     huLimitsBeforeRecenter = .true.
     imposeSaturationLimit = .false.
     imposeRttovHuLimits   = .false.
-    weightRecenter        = 0.0D0  ! means no recentering applied
+    weightRecenter(:)     = -1.0D0 ! means no specified values
     weightRecenterLand    = -1.0D0 ! means same recentering for land as other variables
     numMembersToRecenter  = -1     ! means all members recentered by default
     useOptionTableRecenter = .false.
@@ -317,7 +319,7 @@ contains
       end if
 
       !- Recenter analysis ensemble on supplied analysis
-      if (weightRecenter > 0.0D0 .or. useOptionTableRecenter) then
+      if (any(weightRecenter(:) >= 0.0d0) .or. useOptionTableRecenter) then
         write(*,*) 'epp_postProcess: Recenter analyses on supplied analysis'
         call epp_hybridRecentering(ensembleAnl, weightRecenter, weightRecenterLand, &
                                    useOptionTableRecenter, numMembersToRecenter)
@@ -448,7 +450,7 @@ contains
 
         ! Shift members to have same mean as full ensemble
         call ens_recenter(ensembleAnlSubSample, stateVectorMeanAnl,  &
-                          recenteringCoeff_opt=1.0D0)
+                          recenteringCoeffScalar_opt=1.0D0)
 
         ! Re-compute analysis mean of sub-sampled ensemble
         call ens_computeMean(ensembleAnlSubSample)
@@ -464,7 +466,7 @@ contains
 
         ! Shift members to have same mean as full ensemble
         call ens_recenter(ensembleAnlSubSampleUnPert, stateVectorMeanAnl,  &
-                          recenteringCoeff_opt=1.0D0)
+                          recenteringCoeffScalar_opt=1.0D0)
 
       end if
 
@@ -1356,7 +1358,7 @@ contains
 
     ! Arguments
     type(struct_ens) :: ensembleAnl
-    real(8)          :: weightRecenter
+    real(8)          :: weightRecenter(:)
     real(8)          :: weightRecenterLand
     logical          :: useOptionTableRecenter
     integer          :: numMembersToRecenter
@@ -1371,7 +1373,9 @@ contains
     integer              :: stepIndex, memberIndex, columnIndex
     integer              :: numMembers, numColumns, nulFile, status
     logical              :: recenterAnlFileExists
-    real(8), allocatable :: weightArray(:)
+    real(8), allocatable :: weightArray(:,:)
+    real(8), allocatable :: weightArrayLand(:)
+    real(8)              :: weightFound
     integer, external    :: fnom, fclos
 
     ! check if recentering analysis file exists
@@ -1381,9 +1385,14 @@ contains
       call utl_abort('epp_hybridRecentering: The recentering analysis file does not exist')
     end if
 
+    hco_ens => ens_getHco(ensembleAnl)
+    vco_ens => ens_getVco(ensembleAnl)
     numMembers = ens_getNumMembers(ensembleAnl)
 
-    ! read the optiontable file, if requested
+    allocate( weightArray(maxNumLevels,0:numMembers) )
+    allocate( weightArrayLand(0:numMembers) )
+
+    ! read the optiontable file, if requested    
     if (useOptionTableRecenter) then
       write(*,*) 'epp_hybridRecentering: using optiontable file to specify recentering weights.'
       nulFile = 0
@@ -1394,21 +1403,34 @@ contains
       end if
       call utl_parseColumns(textLine, numColumns)
       if (mmpi_myid==0) write(*,*) 'epp_hybridRecentering: optiontable file has ', numColumns, ' columns.'
-      allocate( weightArray(0:numMembers) )
       rewind(nulFile)
       do memberIndex = 0, numMembers
         read(nulFile,'(a)') textLine
         call utl_parseColumns(textLine, numColumns, stringArray_opt=stringArray)
         if (mmpi_myid==0) write(*,*) memberIndex, (stringArray(columnIndex),columnIndex=1,numColumns)
-        read(stringArray(numColumns),'(f6.3)') weightArray(memberIndex)
-        if (mmpi_myid==0) write(*,*) 'weightArray = ', weightArray(memberIndex)
+        read(stringArray(numColumns),'(f6.3)') weightFound
+        weightArray(:,memberIndex) = weightFound
+        if (mmpi_myid==0) write(*,*) 'weightArray = ', weightArray(1,memberIndex)
       end do
       status = fclos(nulFile)
+    else
+      if (any(weightRecenter(1:vco_getNumLev(vco_ens,'MM')) == -1.d0)) then
+        write(*,*) ''
+        write(*,*) 'Number of weightRecenter coefficients needed = ', vco_getNumLev(vco_ens,'MM')
+        write(*,*) 'Provided values : '
+        write(*,*) weightRecenter(1:vco_getNumLev(vco_ens,'MM'))
+        call utl_abort('epp_hybridRecentering: A valid weightRecenter coefficient was not provided for all the vertical levels')
+      end if
+      do memberIndex = 0, numMembers
+        weightArray(:,memberIndex) = weightRecenter(:)
+      end do
     end if
 
+    do memberIndex = 0, numMembers
+      weightArrayLand(memberIndex) = weightRecenterLand
+    end do
+    
     ! allocate and read in recentering analysis state
-    hco_ens => ens_getHco(ensembleAnl)
-    vco_ens => ens_getVco(ensembleAnl)
     call gsv_allocate( stateVectorRecenterAnl, tim_nstepobsinc, hco_ens, vco_ens, dateStamp_opt=tim_getDateStamp(),  &
                        mpi_local_opt=.true., mpi_distribution_opt='Tiles', &
                        dataKind_opt=4, allocHeightSfc_opt=.false., &
@@ -1421,19 +1443,12 @@ contains
                              stepIndex_opt=stepIndex, containsFullField_opt=.true., &
                              readHeightSfc_opt=.false. )
     end do
-
+    
     ! apply recentering
-    if (useOptionTableRecenter) then
-      call ens_recenter(ensembleAnl, stateVectorRecenterAnl, &
-                        recenteringCoeffArray_opt=weightArray(1:numMembers),  &
-                        numMembersToRecenter_opt=numMembersToRecenter, &
-                        recenteringCoeffLand_opt=weightRecenterLand)
-    else
-      call ens_recenter(ensembleAnl, stateVectorRecenterAnl,  &
-                        recenteringCoeff_opt=weightRecenter,  &
-                        numMembersToRecenter_opt=numMembersToRecenter, &
-                        recenteringCoeffLand_opt=weightRecenterLand)
-    end if
+    call ens_recenter(ensembleAnl, stateVectorRecenterAnl, &
+                      recenteringCoeff_opt=weightArray(:,1:numMembers),  &
+                      numMembersToRecenter_opt=numMembersToRecenter, &
+                      recenteringCoeffLand_opt=weightArrayLand)
 
     call gsv_deallocate(stateVectorRecenterAnl)
     if ( allocated(weightArray) ) deallocate(weightArray)